Moving AnalyticsData into a different package
